narrador_hfy/requirements.txt
# Core
torch~=2.1.0 --index-url https://download.pytorch.org/whl/cu118
TTS>=0.22.0
transformers>=4.35.0
sentencepiece
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
# NLP
spacy>=3.7.0
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.7.0/es_core_news_md-3.7.0-py3-none-any.whl
# Video/Audio
ffmpeg-python>=0.2.0
pillow>=10.1.0
moviepy>=1.0.3
pydub>=0.25.1
# UI
streamlit>=1.28.0
# Utilidades
pyyaml>=6.0
pydantic>=2.5.0
